Move email-related methods to mixin
